requirements.txt
requests>=2.28.0
python-dateutil>=2.8.2
psycopg2-binary>=2.9.3
openai>=0.27.0
regex>=2022.3.15 